src/ml/requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
pydantic==1.10.2
Pillow==9.2.0
torch==1.12.1
transformers==4.21.2
diffusers==0.2.4
transparent-background==1.1.1
opencv-python-headless==4.6.0.66
transparent_background
accelerate